+++ projects/fdc3-web/src/constants.ts
@@ -47,3 +47,5 @@ export const defaultBackoffRetry: Required<BackoffRetryParams> = {
     maxAttempts: 3,
     baseDelay: 250,
 };
+
+export const MS_HOST_MANIFEST_KEY = 'MorganStanley.fdc3-web';